core: harden varint/protobuf parsing and scan-range edge cases; fix numeric random bounds
- reject malformed/overflow varints (Int/Long/UInt + extra-info variants)
- reject negative protobuf length-delimited sizes in skipField/getLength
- handle empty ValueIn safely in key/index scan-range builders
- guard createRanges against empty inputs (avoid divide-by-zero)
- remove unused loop vars (repeat(...)) and keep behavior
- fix random generation ranges for SInt8/SInt16 (include negatives) and UInt8/UInt16 (include max)
- add regression tests for all above in core common tests

+++ core/src/commonMain/kotlin/maryk/core/properties/types/numeric/UInt64.kt
@@ -29,9 +29,9 @@ object UInt64 : UnsignedNumberDescriptor<ULong>(
 
     override fun ofString(value: String) = value.toULong()
     override fun ofDouble(double: Double) = double.toLong().toULong()
-    override fun toDouble(value: ULong) = value.toLong().toDouble()
+    override fun toDouble(value: ULong) = value.toDouble()
     override fun ofInt(int: Int) = int.toULong()
     override fun ofLong(long: Long) = long.toULong()
     override fun createRandom() = Random.nextULong()
-    override fun isOfType(value: Any) = value is UInt64
+    override fun isOfType(value: Any) = value is ULong
 }
